Add header path usage requirements
Instead of include_directories, CMake 3 can use the
target_include_directories command to add required header paths
to a target. This has the advantages of being more specific and
that clients (both internal and external to the project) of the
library only need to link to the target to automatically pick up
the required paths in their compile commands.

Use target_include_directories on quux to specify required header
paths for the Build and Install interfaces. Use PUBLIC/PRIVATE
flag to ensure clients only pick up headers to public interfaces
when using quux.

Retain output of QUUX_INCLUDE_DIR variable in QuuxConfig.cmake
for compatibility and that it may still be useful.